+++ src/components/common/iconButton/CloseButton.tsx
@@ -2,9 +2,9 @@ import React from 'react';
 import { IoClose } from 'react-icons/io5';
 import classnames from 'classnames';
 import './index.scss';
-import { IconAlignEnum, IIconProps } from './types';
+import { IconAlignEnum, IIconButtonProps } from './types';
 
-const CloseButton = ({ onClick, align = IconAlignEnum.right }: IIconProps) => {
+const CloseButton = ({ onClick, align = IconAlignEnum.right }: IIconButtonProps) => {
 	return (
 		<IoClose
 			className={classnames('close-button', { 'align-left': align === IconAlignEnum.left })}
