+++ src/components/common/iconButton/index.scss
@@ -1,3 +1,4 @@
+
 .close-button {
     cursor: pointer;
     position: absolute;
